requirements.txt
gTTS==2.0.3
numpy==1.16.1
opencv-python==4.0.0.21
Pillow==5.4.1
pytesseract==0.2.6